Always parse final "\r\n" when chunking
httpFilterChunkData will always wait for the last "\r\n" in a final chunk
specification, instead of moving on if it is not found. This avoids having a
delayed CRLF pair treated as the beginning of a new request.

This implementation does not allow the sender to include a chunk trailer.